warnings fixed
git-svn-id: file:///afs/psi.ch/project/sls_det_software/svn/slsDetectorSoftware@343 951219d9-93cf-4727-9268-0efd64621fa3

+++ slsDetectorSoftware/slsDetectorAnalysis/angularConversion.h
@@ -152,7 +152,7 @@ class angularConversion : public virtual slsDetectorBase, public angularConversi
       \param bs bin size to be set
       \returns actual bin size
   */
-  double setBinSize(double bs){if (bs>0) nBins=360/bs; return setAngularConversionParameter(BIN_SIZE,bs);};
+  double setBinSize(double bs){if (bs>0) nBins=(int)(360./bs); return setAngularConversionParameter(BIN_SIZE,bs);};
 
   /** 
       get detector bin size
